f11: introduce synchronous model for processor board
The main purpose of synchronous model is to get rid of the latches,
that can't be implemented in the modern FPGA (supporting flips-flops
only) and provide the framework to test synchronous versions of the
dc30x chip models. There is no intention to use this model in FPGA
directly in some projects, the simulation only is supposed.

To achieve these goals:
  - the DMR/SACK logic removed
  - clock generator refactored
  - clock moderators refactored
  - Q-bus logic simplified
  - ODT nBS translation performed at high mclk directly
  - in testbench nBS participates in device decoding

Signed-off-by: 1801BM1 <1801BM1@gmail.com>