f11: fix wrong next microinstruction address
It was found the "divf ACx, ACx" instruction caused the FPP
exception in double precision float mode (see the updated
test.mac for reproduction). The reason was in incorrect
next address bit reset on micro-address 0x008 in the dc303 CS=2.

In this chip the rni signal (read next interrupt - load external
service word with interrupt flags) is generated in the address
range overlapping micro-address correction matrix. Normally, in
the original processor the service word does not cause the next
address bits reset, but in the synchronous models the service
word is loaded half-cycle later and D-register occasional value
caused the next-address bit 2 reset, resulting in wrong microcode
sequence and instruction failure.

As the PDP-11 instruction decode matrix in the Control Chip
is not intended to decode the interrupt service word we can mask
out the next address reset signals if rni is active.

Signed-off-by: 1801BM1 <1801BM1@gmail.com>