cpu: support simulation with Icarus Verilog
Iarus Verilog is an implementation of the Verilog hardware description
language compiler that generates netlists in the desired format (EDIF).
It upports the 1995, 2001 and 2005 versions of the Verilog standard,
portions of SystemVerilog, and some extensions.

The configuration files and running scripts are added in synchronous
models for simulation with Icarus Verilog.

Signed-off-by: Yuriy Shestakov <yshestakov@gmail.com>