vm2: fix shadow PSW and PC update in halt mode
When processor goes to halt mode the shadow PSW and PC
registers are held to keep the return vector to user mode.
The exception handler in halt mode can take a long time to
be executed (as example, handling the float point instruction
emulation) and interrupts can be enabled in halt mode by
resetting PSW[7] bit. This unlocks the update of shadow PSW
and PC and these registers are saved on stack on interrupt
occurrence in halt mode.

Once the PSW[7] bit changes its state from one to zero the
shadow PSw and PC registers are not updated on this first
PSW write. If there is the pending interrupt the non-updated
valued from shadow registers will be saved to stack, and there
will be incorrect return to the halt mode program.

The fix checks if the new PSW[7] and PSW[8] values enable
the shadow registers update and allows their update at once.

Reported-by: xolod79
Signed-off-by: 1801BM1 <1801BM1@gmail.com>